# cmd/api/docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      err:
        type: string
      hint:
        type: string
      status:
        type: string
    type: object
...
  models.OrdersListing:
    properties:
      addressId:
        type: integer
      bookId:
        type: integer
      deliveryStatus:
        type: boolean
      id:
        type: integer
      paymentId:
        type: string
      paymentMethodId:
        type: integer
      paymentStatus:
        type: string
      quantity:
        type: integer
      totalPrice:
        type: number
      userId:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /admin/allorders:
    get:
      description: Retrieve all orders based on a filter.
      parameters:
      - description: Filter parameter
        in: query
        name: filter
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            items:
              items:
                $ref: '#/definitions/models.OrdersListing'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all orders
      tags:
      - order
...
  /user/addorder:
    post:
      description: Place an order for a user with specified details.
      parameters:
      - description: User ID
        in: query
        name: userid
        required: true
        type: integer
      - description: Payment Method ID
        in: query
        name: paymentmethoadid
        required: true
        type: integer
      - description: Address ID
        in: query
        name: adressid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order placed
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Add an order
      tags:
      - order
...
  /user/cancelorder:
    delete:
      description: Cancel an order for a user with specified details.
      parameters:
      - description: User ID
        in: query
        name: userid
        required: true
        type: integer
      - description: Order ID
        in: query
        name: orderid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order cancelled
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Cancel an order
      tags:
      - order
...
  /user/getorder:
    get:
      description: Retrieve details of a specific order for a user.
      parameters:
      - description: User ID
        in: query
        name: userid
        required: true
        type: integer
      - description: Order ID
        in: query
        name: orderid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            $ref: '#/definitions/models.OrdersListing'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get order details
      tags:
      - order
...
  /user/listorder:
    get:
      description: Retrieve all orders for a user.
      parameters:
      - description: User ID
        in: query
        name: userid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            items:
              $ref: '#/definitions/models.OrdersListing'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: list a users orders
      tags:
      - order
...
swagger: "2.0"